+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-12-04 15:13:10
+
+### Changed Files
+- `src/features/admin/components/AdminControlCenter.tsx`
+
+### Statistics
+```
+ 1 file changed, 8 insertions(+), 1 deletion(-)
+```
+
 ## [Checkpoint] - 2025-12-04 14:24:26
 
 ### Changed Files